pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "datarobot-genai"
version = "0.1.17"
description = "Generic helpers for GenAI"
readme = "README.md"
requires-python = ">=3.10, <3.13"
license = { text = "Apache-2.0" }
authors = [{ name = "DataRobot, Inc." }]
dependencies = [
  "requests>=2.32.4",
  "datarobot>=3.9.1",
  "datarobot-predict>=1.13.2",
  "openai>=1.76.2",
  "pandas>=2.2.3",
  "ragas>=0.3.8",
]

[project.urls]
Homepage = "https://github.com/datarobot-oss/datarobot-genai"

[project.optional-dependencies]
crewai = [
  "crewai>=0.193.2",
  "crewai-tools>=0.69.0",
  "opentelemetry-instrumentation-crewai>=0.40.5",
  "pybase64>=1.4.2",
  "pylance<=0.38.1",
]
langgraph = [
  "langgraph>=0.4.10",
  "langgraph-prebuilt>=0.2.3",
]
llamaindex = [
  "llama-index>=0.13.6,<0.14.0",
  "llama-index-core>=0.13.6,<0.14.0",
  "llama-index-llms-langchain>=0.6.1",
  "llama-index-llms-litellm>=0.4.1",
  "llama-index-llms-openai>=0.3.38",
  "opentelemetry-instrumentation-llamaindex>=0.40.5",
  "pypdf>=6.0.0",
]
nat = [
  "nvidia-nat==1.3.0; python_version >= '3.11'",
  "nvidia-nat-opentelemetry==1.3.0; python_version >= '3.11'",
  "nvidia-nat-crewai==1.3.0; python_version >= '3.11'",
  "nvidia-nat-langchain==1.3.0; python_version >= '3.11'",
]
pydanticai = [
  "logfire>=4.6.0",
  "pydantic-ai>=1.0.5",
]
drmcp = [
  "datarobot-asgi-middleware>=0.2.0",
  "python-dotenv>=1.1.0",
  "boto3>=1.34.0",
  "httpx>=0.28.1",
  "pydantic>=2.6.1",
  "pydantic-settings>=2.1.0",
  "opentelemetry-api>=1.22.0",
  "opentelemetry-sdk>=1.22.0",
  "opentelemetry-instrumentation-requests>=0.43b0",
  "opentelemetry-instrumentation-aiohttp-client>=0.43b0",
  "opentelemetry-instrumentation-httpx>=0.43b0",
  "opentelemetry-exporter-otlp>=1.22.0",
  "opentelemetry-exporter-otlp-proto-http>=1.22.0",
  "aiohttp>=3.9.0",
  "aiohttp-retry>=2.8.3",
  "aiosignal>=1.3.1",
  "fastmcp==2.12.2",
]

[tool.hatch.build.targets.wheel]
packages = [
  "src/datarobot_genai",
]

[tool.hatch.build.targets.sdist]
include = [
  "src/datarobot_genai",
  "pyproject.toml",
  "README.md",
  "LICENSE",
  "AUTHORS",
]

[tool.uv]
package = true
override-dependencies = [
  "ragas>=0.3.8",
]

[dependency-groups]
dev = [
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "pytest",
  "pytest-cov",
  "pre-commit",
  "twine>=5.0.0",
  "pytest-asyncio>=0.23.5",
  "responses>=0.25.8",
]


[tool.mypy]
disallow_untyped_defs = true
disable_error_code = ["import-untyped"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",  # Pyflakes
    "F",  # pycodestyle
    "I",  # isort
    "N",  # pep8-naming
    "D",  # pydocstyle
    "PL", # pylint
    "UP", # pyupgrade
    "INP", # check missing __init__.py
    "B006", # mutable-argument-default
]
extend-ignore = [
    # D100-D106: Ignore undocumented module, class, method, function & inner class
    "D100", "D101", "D102", "D103", "D104", "D105", "D106",
    # 1 blank line required between summary line and description
    "D205",
    # Too many arguments
    "PLR0913",
    # Too many statements
    "PLR0915",
    # Too many branches
    "PLR0912",
]
fixable = [
    "I",    # isort
    "D",    # pydocstyle
    "UP",   # pyupgrade
    "F401", # Imported but unused
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "D104", # Missing docstring in public package
    "F401"  # Imported but unused
]
"scripts/*.py" = [
    "INP001", # implicit-namespace-package: allow scripts without package semantics
    "E402",   # module-import-not-at-top-of-file: allow local imports in smoke script
    "I001",   # isort import order within functions
    "PLC0415",# pylint import-outside-toplevel for controlled local imports
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.pylint]
allow-magic-value-types = ["int", "str", "bytes"]
max-args = 8

[tool.pytest.ini_options]
filterwarnings = []

[tool.coverage.run]
source = ["datarobot_genai"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError",
    "pass",
]
fail_under = 80
